backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
googletrans==4.0.0rc1
cohere==5.15.0
python-dotenv==1.0.1
gunicorn==21.2.0